Initial configuration for bag recording and foxglove bridge (#1)
This PR provides an initial configuration/setup for the ros2 bag
recording and foxglove bridge from a docker container.

- Two types of scripts directories are introduced `host_tools` and
`container_tools`. As it names implies `host_tools`provides scripts that
are intended for host usage, i.e outside docker container. Moreover
`container_tools`, offer scripts that are needed inside the docker
container. The summary of the scripts states as follows:
  - `host_tools`:
- `upload_docker_image_to_vehicle.sh`: Receive a docker image as input,
to save it with `docker save` and send it over the vehicle uisng `scp`.
Docker cmd are used for bash autocomplate, thus after doing "tab tab"
method, the only docker images available in your system will be
displayed
- `install_tool.sh`: Install the upload docker script and
autocompletion.
  - `container_tools`:
- `foxglove_bridge.sh`: Runs foxglove_bridge launch file with some
arguments
- `record_rosbag.sh`: Runs `ros2 bag record` cmd to allow `mcap` bag
format. Optionally a `.txt` file can be passed as argument to define
record-only topics.
   
- Docker: The docker build implements dev and runtime build stages.
Majority of the install relies on `base` stage. The default command for
both sages are `bash` and `record_rosbag` respectively
- `dev.sh` and `runtime.sh` scripts are provided to ease development and
testing. By default BOTH scripts create a `rosbag` dir if it does not
exists and mount it to the dockercontainer